Feature/sqlite user (#16)
* Add sqlite UserService
New UserService that uses Sqlite database as a storage.
Had to change the `api.User.ID` from `uint64` to `int64`.
Apparently, databases don't like `uint64`.
Shouldn't be a problem though. The `int64` is still plenty.
And the ID is generated with `rand.Int63` so there will be not negative IDs.

* Add DBOptions type for database configuration.
The database connection string can be passed from the main through the
`ApiServer.Option` to the `UserService`.

* Add servers startup error handling
If either API or Web server fails to start for some reason main will gracefully exit.